Adding uns_functions module to flattener_mods. Moving over uns functions, as well as section of uns code from main, from flattener to uns_functions.
