+++ scripts/flattener_mods/__init__.py
@@ -1,2 +1,3 @@
 from flattener_mods.gather import gather_rawmatrices, gather_objects, gather_metdata, gather_pooled_metadata, get_value
 from flattener_mods.constants import UNREPORTED_VALUE, SCHEMA_VERSION, MTX_DIR, REF_FILES, CELL_METADATA, DATASET_METADATA, ANNOT_FIELDS, ANTIBODY_METADATA, PROP_MAP
+from flattener_mods.uns_functions import colors_check, check_not_empty, copy_over_uns, process_spatial
